migrate memoDerivation
